convert/pyproject.toml
[project]
name = "rapid-ocr-rknn-convert"
version = "0.1.0"
description = "Utility to convert PaddleOCR models from ONNX to RKNN."
requires-python = ">=3.11,<3.13"
dependencies = [
    "kagglehub>=1.0.0",
    "onnxruntime>=1.24.4",
    "opencv-python>=4.11.0.86",
    "rapidocr>=3.8.1",
    "rknn-toolkit2>=2.3.2",
    "setuptools<81",
    "typed-argparse>=0.3.1",
]
